fix and add some test
