+++ tests/client/src/components/QA/QA.test.js
@@ -3,7 +3,11 @@ import { shallow, mount } from 'enzyme';
 
 
 import QA from '../../../../../client/src/components/QA/QA.jsx';
+import Question from '../../../../../client/src/components/QA/Question.jsx';
 import QuestionList from '../../../../../client/src/components/QA/QuestionsList.jsx';
+import SearchQuestion from '../../../../../client/src/components/QA/SearchQuestion.jsx';
+import AnswersList from '../../../../../client/src/components/QA/AnswersList.jsx';
+import AddingForm from '../../../../../client/src/components/QA/AddingForm.jsx';
 import {productData, questionsData, answersData} from './data.js';
 
 // ==== Test Template ====
@@ -16,23 +20,32 @@ describe('<QA testing />', () => {
     const text = wrapper.find('h3').text();
     expect(text).toBe('QUESTIONS & ANSWERS');
     expect(wrapper).toMatchSnapshot();
-    const button = wrapper.find('button').length;
-    expect(button).toBe(1);
   });
 
-  
-});
-
-describe ('Render questions and answer', () => {
-  const wrapper = mount(<QuestionList questions={questionsData} />);
-  test('should render question body', () => {
+  test('renders Question List', () => {
+    const wrapper = shallow(<QuestionList questions={questionsData} />);
     expect(wrapper).toMatchSnapshot();
-    expect(wrapper.find('.qa-question')).toHaveLength(2);
   });
 
-  test('should add 2 questions when click more answered question button', () => {
-    expect(wrapper.find('table.qa-question-table')).toHaveLength(2);
-    wrapper.find('.button#qa-more-question').simulate('click');
-    expect(wrapper.find('table.qa-question-table')).toHaveLength(4);
-  });  
+  test('renders Question', () => {
+    const wrapper = shallow(<Question questions={questionsData} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  test('renders Search Question', () => {
+    const wrapper = shallow(<SearchQuestion questions={questionsData} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  test('renders Answers List', () => {
+    const wrapper = shallow(<AnswersList answers={answersData} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  test('renders Adding form', () => {
+    const wrapper = shallow(<AddingForm productId='36300' />);
+    expect(wrapper).toMatchSnapshot();
+  });
+
 });
+
